Ensure new conversation deselects nav (#267)
