+++ app/views/messages/_main_column.html.erb
@@ -65,6 +65,7 @@
     <%= render 'messages/main_column/assistant_header' %>
 
     <%= link_to new_assistant_message_path(@assistant),
+        target: "_top",
         class: "block cursor-pointer inline-flex mr-3 px-2 py-2 group",
         data: {
           role: "new",
@@ -93,7 +94,7 @@
 
       <%= render 'messages/main_column/assistant_header' %>
 
-      <%= link_to new_assistant_message_path(@assistant), class: "block cursor-pointer inline-flex mr-1 px-2 group", data: { role: "new" } do %>
+      <%= link_to new_assistant_message_path(@assistant), target: "_top", class: "block cursor-pointer inline-flex mr-1 px-2 group", data: { role: "new" } do %>
         <%= icon "pencil-square", variant: :outline, size: 22, class: 'stroke-2 text-gray-950 dark:text-gray-100', title: "New chat", tooltip: :left %>
       <% end %>
     </div>
