+++ test/system/conversations/messages/auto_scroll_test.rb
@@ -63,4 +63,19 @@ class ConversationMessagesAutoScrollTest < ApplicationSystemTestCase
       end
     end
   end
+
+  test "when conversation is scrolled to the bottom, when the browser resizes it auto-scrolls to stay at the bottom" do
+    assert_stays_at_bottom do
+      resize_browser_to(1400, 700)
+    end
+  end
+
+  test "when conversation is NOT scrolled to the bottom, when the browser resizes it DOES NOT auto-scroll so what scrolled to stays visible" do
+    assert_at_bottom
+    assert_scrolled_up { scroll_to find_messages.second }
+
+    assert_did_not_scroll do
+      resize_browser_to(1400, 700)
+    end
+  end
 end
